refactor: 같은 패키지 내에 날짜 도메인 관리
