--- src/main/java/christmas/domain/Calendar.java
@@ -1,4 +1,0 @@
-package christmas.domain;
-
-public class Calendar {
-}
