+++ src/main/java/christmas/domain/calendar/Calendar.java
@@ -1,0 +1,4 @@
+package christmas.domain.calendar;
+
+public class Calendar {
+}
