+++ src/main/java/christmas/domain/discount/WeekendDiscountPolicy.java
@@ -1,6 +1,6 @@
 package christmas.domain.discount;
 
-import christmas.domain.DayChecker;
+import christmas.domain.calendar.DayChecker;
 import christmas.domain.FoodCategory;
 import christmas.domain.Order;
 import christmas.domain.Orders;
